+++ tests/ZestInstTest.php
@@ -92,8 +92,11 @@ class ZestInstTest extends \PHPUnit\Framework\TestCase {
 	public function testGetElementByClassName() {
 		$doc = self::loadHTML( __DIR__ . "/index.html" );
 		$doc->documentElement->setAttribute( 'class', "testGetElementByClassName\t" );
-		$this->assertSame( $doc->documentElement, TestingAccessWrapper::newFromObject( new ZestInst )
-			->getElementsByClassName( $doc, 'testGetElementByClassName' )->item( 0 ) );
+		$this->assertSame(
+			$doc->documentElement,
+			TestingAccessWrapper::newFromObject( new ZestInst )
+			->getElementsByClassName( $doc, 'testGetElementByClassName', [] )->item( 0 )
+		);
 	}
 
 	public static function toXPath( $node ) {